Refactor to reduce loops
